requirements.txt
pkg-resources==0.0.0
pyfiglet==0.8.post1
python-nmap==0.6.1
tabulate==0.8.6
